MAINT: housecleaning for traversal loop setup and implementations
